+++ main.py
@@ -1,47 +1,8 @@
-import logging, argparse, datetime, pickle
+import logging, argparse, datetime
+import src.seleniumScraper as scraper
+from src.lobbyingDataPage import PageFactory
+from src.sql_manager import execute_query
 
-import src.settings as settings
-import src.lobbyingDataPage as ldp
-import src.lobbyingScraper as ls
-import src.utils as utils
-
-
-########################
-# COMMAND LINE METHODS #
-########################
-# -a --all
-def pull_and_process_all_disclosure_urls():
-    current_year = datetime.datetime.now().year
-    scrape_disclosure_urls(end_year=current_year)
-    process_disclosure_urls()
-
-# -r --recent
-def pull_and_process_recent_disclosure_urls():
-    current_year = datetime.datetime.now().year
-    scrape_disclosure_urls(current_year-1, current_year)
-    process_disclosure_urls()
-
-###############################
-# Command line Helper Methods #
-###############################
-
-def scrape_disclosure_urls(start_year=2005, end_year=2023):
-    for year in range(start_year, end_year + 1):
-        logging.info(f"Pulling disclosure report urls from {year}")
-        ls.get_disclosures_by_year(year)
-
-def process_disclosure_urls():
-    all_disclosures = set(utils.execute_query("SELECT url FROM disclosure_urls;"))
-    processed_disclosures = set(utils.execute_query("SELECT url FROM headers;"))
-    new_disclosures = all_disclosures - processed_disclosures
-    print(f"Processing {len(new_disclosures)} disclosure url's")
-    for disclosure_url in new_disclosures:
-        ldp.PageFactory(disclosure_url[0]).save()
-
-
-###################
-# Argument Parser #
-###################
 
 def parse_arguments():
     verbosity_levels = [logging.WARNING, logging.INFO, logging.DEBUG]
@@ -63,11 +24,11 @@ def parse_arguments():
     logging.basicConfig(level=verbosity_levels[args.verbose])
 
     if args.all:
-        pull_and_process_all_disclosure_urls()
+        scraper.pull_and_process_all_disclosure_urls()
     elif args.recent:
-        pull_and_process_recent_disclosure_urls()
+        scraper.pull_and_process_recent_disclosure_urls()
     elif args.process:
-        process_disclosure_urls()
+        scraper.process_disclosure_urls()
     else:
         parser.print_help()
 
@@ -75,4 +36,44 @@ def parse_arguments():
 # Main
 if __name__ == "__main__":
     parse_arguments()
+
+
+########################
+# COMMAND LINE METHODS #
+########################
+# -a --all
+def pull_and_process_all_disclosure_urls():
+    current_year = datetime.datetime.now().year
+    scrape_disclosure_urls(end_year=current_year)
+    process_disclosure_urls()
+
+# -r --recent
+def pull_and_process_recent_disclosure_urls():
+    current_year = datetime.datetime.now().year
+    scrape_disclosure_urls(current_year-1, current_year)
+    process_disclosure_urls()
+
+
+
+###############################
+# Command line Helper Methods #
+###############################
+
+def scrape_disclosure_urls(start_year=2005, end_year=2023):
+    driver = scraper.create_driver()
+    for year in range(start_year, end_year + 1):
+        logging.info(f"Pulling disclosure report urls from {year}")
+        scraper.get_disclosures_by_year(year,driver)
+    driver.close()
+
+def process_disclosure_urls():
+    driver = scraper.create_driver()
+    all_disclosures = set(execute_query("SELECT url FROM disclosure_urls;"))
+    processed_disclosures = set(execute_query("SELECT url FROM headers;"))
+    new_disclosures = all_disclosures - processed_disclosures
+    print(f"Processing {len(new_disclosures)} disclosure url's")
+    for disclosure_url in new_disclosures:
+        PageFactory(disclosure_url[0], driver).save()
+    driver.close()
+
